requirements.txt
tensorflow
streamlit
matplotlib
pillow
plotly
opencv-python
keras-squeezenet
tensorflow-addons
